==> stimuli/stimulus_generation/1_concreteness_manipulation.Rmd <==
---
title: "FakeDiamond: concreteness manipulation"
author: "Ryan Law"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(LexOPS)
library(tidyverse)
```

## Generate a list of single words 

As a first step in our stimulus creation, we focus on our concreteness manipulation. 

We generate a list of concrete and abstract words by sampling from the two local maxima in the concrete rating distribution, ranging from 1 to 5, from the Brysbaert norms (Brysbaert et al., 2014). Concrete words were drawn from the range of 1.5-2.5, while abstract words were drawn from between 4-5. 

All pairs of concrete-abstract words are matched on several psycholinguistic variables known to influence lexical and conceptual access:
- part of speech: all words are filtered to be nouns
- proportion known: at least 90% of raters know these words
- word length: these words are 3-10 letters long; each concreteness pair is allowed to differ within the pair 2 letters
- single word frequency in Zipf (frequency per billion words; variance +-0.3)
- age of acquisition (variance +-2 years)
- number of syllables (+-2)
- number of phonemes (+-2)
- lexical reaction time (+-50 ms)
- valence (+-1)

```{r create list of words}
stim <- LexOPS::lexops %>%
	subset(PK.Brysbaert >= 0.9 & PK.Brysbaert <= 1) %>%
	subset(Length >= 3 & Length <= 10) %>%
	subset(PoS.SUBTLEX_UK %in% c("noun")) %>%
  split_by(CNC.Brysbaert, 1.5:2.5 ~ 4:5) %>%
	control_for(Zipf.SUBTLEX_UK, -0.3:0.3) %>%
	control_for(AoA.BrysbaertBiemiller, -2:2) %>%
  control_for(Syllables.CMU, -2:2) %>%
  control_for(Phonemes.CMU, -2:2) %>%
	control_for(RT.BLP, -50:50) %>%
  control_for(VAL.Glasgow_Norms, -1:1) %>%
	control_for(Length, -2:2) %>%
	generate("all", "balanced", seed = 42)
stim_long <- long_format(stim)
rep_str = c('A1'='Abstract','A2'='Concrete')
stim_long$condition <- str_replace_all(stim_long$condition, rep_str)
# write.csv(stim_long, sprintf("U:\\ownCloud\\projects\\FakeDiamond\\stimuli\\concreteness_denotation\\1_word_list_seed42_%s.csv", Sys.Date()), row.names=FALSE)
write.csv(stim_long, "U:\\ownCloud\\projects\\FakeDiamond\\stimuli\\concreteness_denotation\\1_concreteness_list.csv", row.names=FALSE)
stim_long
```

## Examine distribution

```{r visualise design, echo=FALSE}
plot_design(stim_long)
```
