==> stimuli/stimulus_generation/2_subselect_nouns.ipynb <==
# Subselecting nouns and visualise psycholinguistic properties

We now have a list of concrete-abstract word pairs generated using LexOPS (Taylor et al., 2020) and stored as `word_list_seed42_2023-04-14.csv`. In this notebook, we want to (1) remove some nouns because they are not appropriate for an experiment and (2) visualise the stimulus psycholinguistic properties.

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from datetime import date

dir = 'U:\ownCloud\projects\FakeDiamond\stimuli'

First read in our word list generated with our RMarkdown script and do some tidying up.

# tidy up our pre-selected concrete-abstract words 
word_list = pd.read_csv(os.path.join(dir, 'concreteness_denotation/1_concreteness_list.csv')) 

# change all strings to lowercase
word_list = word_list.applymap(lambda s: s.lower() if type(s) == str else s) 

# rename columns
word_list.rename(columns={'string':'word','item_nr':'set_nr','condition':'concreteness','CNC.Brysbaert':'CNC_M','Zipf.SUBTLEX_UK':'Zipf','AoA.BrysbaertBiemiller':'AoA','Syllables.CMU':'Syllables','Phonemes.CMU':'Phonemes','VAL.Glasgow_Norms':'Valence','RT.BLP':'RT'}, inplace=True)

# drop irrelevant column
word_list.drop(labels=['match_null'], axis=1, inplace=True) 

Next, we want to import the *standard deviation* of concreteness ratings, also from Brysbaert and colleagues (2014), as well as the imageability of these words (i.e., how easy raters find it to visualise these words; Scott et al., 2019). 

# extract concreteness rating standard deviation from Brysbaert et al. (2014)
brysbaert_ratings = pd.read_csv(os.path.join(dir, 'resources', 'brysbaert_etal_2014.csv'))
brysbaert_ratings.rename(columns={'Word': 'word'}, inplace=True)
brysbaert_ratings.set_index('word', inplace=True)
word_list['CNC_SD'] = ''
word_list.set_index('word', inplace=True)
word_list['CNC_SD'] = word_list.index.map(brysbaert_ratings['Conc_SD'])
word_list.reset_index(inplace=True)

# add imageability ratings from Scott et al. (2019)
scott_ratings = pd.read_csv(os.path.join(dir, 'resources', 'scott_etal_2019.csv'))
scott_ratings.rename(columns={'Words': 'word'}, inplace=True)
scott_ratings.set_index('word', inplace=True)
word_list['IMAG'] = ''
word_list.set_index('word', inplace=True)
word_list['IMAG'] = word_list.index.map(scott_ratings['IMAG'])
word_list.reset_index(inplace=True)
word_list.rename(columns={'word': 'noun'}, inplace=True)
word_list['IMAG'] = pd.to_numeric(word_list['IMAG'])
word_list

Some nouns are not the most appropriate for a cognitive neuroscience experiment (based on intuitive judgement). Some nouns are mostly used in North America and might not be the most familiar to native speakers from our participant pool, who are more likely to be native British English speakers. As a quick diagnostic, we also check whether the nouns go well with the adjective `fake` (e.g., `fake nay`). So, I subselected them, along with their pair (to keep the psycholinguistic property matching), out by hand using Excel and keep a record of them here. 

nouns_to_be_removed = ['bland', 'slave', 'thrifty', 'vulgar', 'quantum', 'cervix', 'anus', 'doubt', 'meek', 'morsel', 'pimp', 
                       'innate', 'maggot', 'premium', 'comet', 'uplift', 'gag', 'urine', 'gripe', 'dung', 'hooker', 'fraud',
                       'farewell','delight','sperm','midst','kind','sane','thanks','mistress','stroke','awe']

# remove unwanted nouns
nouns_cleaned = word_list[~word_list.noun.isin(nouns_to_be_removed)]

# remove counterparts of unwanted nouns
for set_nr in range(1, len(nouns_cleaned.set_nr.unique())+1):
    if len(nouns_cleaned[nouns_cleaned.set_nr==set_nr]) < 2:
        nouns_cleaned = nouns_cleaned[nouns_cleaned.set_nr != set_nr]

# reset set number, reset index, sort dataframe
nouns_cleaned.set_nr = np.repeat(range(1, int(len(nouns_cleaned)/2)+1), 2)
nouns_cleaned.reset_index(inplace=True, drop=True)
nouns_cleaned.sort_values(by=['set_nr','concreteness'], inplace=True, ascending=[True,False])
nouns_cleaned

Save to disk for phrase creation

nouns_cleaned.to_csv(os.path.join(dir, 'concreteness_denotation/2_nouns_cleaned.csv'), index=False) # save to disk
