==> stimuli/stimulus_generation/4_calculate_M_SD.ipynb <==
# Calculate word and phrasal level characteristics for reporting in paper

import pandas as pd

stim = pd.read_excel('3_phrases_2023-07-11.xlsx', index_col=0)
stim = stim[stim.denotation == 'baseline']

Look at word-level charateristics

columns_to_summarize = ['Length', 'Zipf', 'AoA','CNC_M','RT','IMAG','Valence']
aggregations = {col: ['mean', 'std'] for col in columns_to_summarize}

summary = stim.groupby('concreteness').agg(aggregations)
summary.columns = ['_'.join(col) for col in summary.columns]  # Flatten MultiIndex columns
summary = summary.reset_index()

print(summary)


Look at phrasal level statistics

stim = pd.read_excel('3_phrases_2023-07-11.xlsx', index_col=0)
stim = stim[stim.denotation != 'baseline']
stim

columns_to_summarize = ['frequency_dep', 'frequency_seq', 'frequency_dep_log','frequency_seq_log']
aggregations = {col: ['mean', 'std'] for col in columns_to_summarize}

summary = stim.groupby(['concreteness','denotation']).agg(aggregations)
summary.columns = ['_'.join(col) for col in summary.columns]  # Flatten MultiIndex columns
summary = summary.reset_index()

print(summary)


measures = ['frequency_dep', 'frequency_seq']
final_table = pd.DataFrame(index=measures)

for index, row in summary.iterrows():
    col_name = f"{row['concreteness']}-{row['denotation']}"
    
    for measure in measures:
        mean = row[f'{measure}_mean']
        std = row[f'{measure}_std']
        final_table.loc[measure, col_name] = f"{mean:.2e} ({std:.2e})"

column_order = [
    'concrete-subsective',
    'concrete-privative',
    'abstract-subsective',
    'abstract-privative'
]
final_table = final_table.reindex(columns=column_order)


print("Descriptive Statistics for Stimulus Frequencies")
print(final_table.to_string())
final_table
