==> stimuli/stimulus_generation/5_make_stimuli_file_for_meg.ipynb <==
This script takes stimulus materials created for our three experimental manipulations and put them into one stimulus file, to be taken as input to the stimulus delivery script.

import numpy as np
import pandas as pd
from random import sample

import matplotlib.pyplot as plt
import seaborn as sns

from scipy import stats

stimuli_concreteness_denotation = pd.read_excel('concreteness_denotation/stimuli_concreteness_denotation_complete.xlsx')
stimuli_specificity = pd.read_excel('specificity/stimuli_specificity_complete.xlsx')

# create experiment ID (concreteness_denotation, specificity) for ease of subsetting during analysis
experiment = ['concreteness_denotation'] * len(stimuli_concreteness_denotation) + ['specificity'] * len(stimuli_specificity)

# maintain set-hood information across experiments
set_nr = np.repeat(range(1,101),6).tolist() + np.repeat(range(1,101),3).tolist()

# create item-specific ID for ease of indexing during analysis
item_nr = list(range(1,len(stimuli_concreteness_denotation)+len(stimuli_specificity)+1))

# create condition name
condition_name = (stimuli_concreteness_denotation.concreteness + '_' + stimuli_concreteness_denotation.denotation).tolist() + stimuli_specificity.specificity.tolist()

# create word 1 and word 2
word1 = stimuli_concreteness_denotation.adjective.tolist() + stimuli_specificity.word1.tolist()
word2 = stimuli_concreteness_denotation.noun.tolist() + stimuli_specificity.word2.tolist()

# include probe and answer
probe = stimuli_concreteness_denotation.probe.tolist() + stimuli_specificity.probe.tolist()
answer = stimuli_concreteness_denotation.answer.tolist() + stimuli_specificity.answer.tolist()

# put into a df
stimuli = pd.DataFrame()
stimuli['experiment'] = experiment
stimuli['set_nr'] = set_nr
stimuli['item_nr'] = item_nr
stimuli['condition_name'] = condition_name
stimuli['word1'] = word1
stimuli['word2'] = word2
stimuli['probe'] = probe
stimuli['answer'] = answer
stimuli['trigger_word1'] = ''
stimuli['trigger_word2'] = ''

# create trigger (concrete: 100; abstract: 200; subsective: 10; privative: 20; baseline: 30; low: 70; mid: 80; high: 90; word1: 1; word2: 2)
# they add up to make the trigger for each condition (e.g., concrete_subsective_word1 is 100+10+1=111; abstract_baseline_word2 is 200+30+2=232)
stimuli.loc[stimuli.condition_name=='concrete_subsective','trigger_word1'] = 111
stimuli.loc[stimuli.condition_name=='concrete_privative','trigger_word1'] = 121
stimuli.loc[stimuli.condition_name=='concrete_baseline','trigger_word1'] = 131
stimuli.loc[stimuli.condition_name=='abstract_subsective','trigger_word1'] = 211
stimuli.loc[stimuli.condition_name=='abstract_privative','trigger_word1'] = 221
stimuli.loc[stimuli.condition_name=='abstract_baseline','trigger_word1'] = 231
stimuli.loc[stimuli.condition_name=='concrete_subsective','trigger_word2'] = 112
stimuli.loc[stimuli.condition_name=='concrete_privative','trigger_word2'] = 122
stimuli.loc[stimuli.condition_name=='concrete_baseline','trigger_word2'] = 132
stimuli.loc[stimuli.condition_name=='abstract_subsective','trigger_word2'] = 212
stimuli.loc[stimuli.condition_name=='abstract_privative','trigger_word2'] = 222
stimuli.loc[stimuli.condition_name=='abstract_baseline','trigger_word2'] = 232
stimuli.loc[stimuli.condition_name=='low','trigger_word1'] = 71
stimuli.loc[stimuli.condition_name=='mid','trigger_word1'] = 81
stimuli.loc[stimuli.condition_name=='high','trigger_word1'] = 91
stimuli.loc[stimuli.condition_name=='low','trigger_word2'] = 72
stimuli.loc[stimuli.condition_name=='mid','trigger_word2'] = 82
stimuli.loc[stimuli.condition_name=='high','trigger_word2'] = 92
stimuli

stimuli.to_csv('../scripts/stimulus_presentation/stimuli_test.csv', index=False)

It's useful to have a small set of items as a debug stimulus set:

# stimuli_test = stimuli.sample(frac=0.007, random_state=70)
# stimuli_test.to_csv('../scripts/stimulus_presentation/stimuli_debug.csv', index=False)

After testing the script in the scanner, a total of 1182 trials, with a long enough inter-trial interval for cortical activations to return to baseline (\~2s), the experiment would be too long (\~1.5hrs effective recording time; longer with setting up and brief breaks between blocks). We decided to reduce the number of trials per condition to 100, yielding 900 trials total (effective recording time \~43 minutes). Below we document how we reduce the number of trials.
